Make SpecializedSingleQubitControlledGate a bloq.
